Don't compute transform footprints if logging is not enabled.
I would expect the computation of these to get optimised out, but the optimiser
(GCC and CLANG) can't see that *no* thread uses the computed value or the
changed data structure. I don't blame them for erring on the side of caution
(because I'm changing a shared memory structure), but we lose nothing by making
it a little easier for the compiler to read.